+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [2.0.0] 2021-01-10
+### Changed
+- Inject button for the registration nodes (input removed).
+- Outputs for the registration nodes to signal success or errors.
+- Sender node: input message interface extended for future data. Outputs to singnal success or errors.
+- Receiver node: Output message content extended. Outputs to singnal success or errors.
+
+### Added
+- Example flows
+
 ## [1.3.1] 2021-01-04
 ### Fixed
 - Message receiver management.